Move Turing OF headers out of third_party
